src/robot_control.py
from robodk import *

class RobotController:
    def __init__(self, robot_item):
        self.robot = robot_item

    def move_to_position(self, target_position, speed=100):
    #"""
    #Move the robotic arm to the specified target position.
    #This should be within the robot's joint limits.
    #"""
    # Check if the target position is within the robot's joint limits
    joint_limits = self.robot.JointLimits()  # Get the joint limits
    for i, pos in enumerate(target_position):
        if pos < joint_limits[i][0] or pos > joint_limits[i][1]:
            raise ValueError(f"Target position {pos} for joint {i} is outside limits.")

    # Move the robot to the target position
    self.robot.MoveJ(target_position, speed)

    def pick_object(self, bbox):
        """
        Move the robot to the position above the object and perform the pick action.

        :param bbox: A tuple containing the bounding box coordinates (x, y, w, h) of the object.
        """
        x, y, w, h = bbox
        pick_position = [x + w / 2, y + h / 2, 100]  # Adjust the Z value as needed for your setup

        # Move to the pick position
        self.move_to_position(pick_position)

        # Perform the pick action (e.g., close the gripper)
        self.gripper_action('close')

    def place_object(self, target_position):
        """
        Move the robot to the target position and perform the place action.

        :param target_position: A list or array containing the target position (x, y, z) where the object will be placed.
        """
        # Move to the place position
        self.move_to_position(target_position)

        # Perform the place action (e.g., open the gripper)
        self.gripper_action('open')

    def gripper_action(self, action):
        """
        Control the gripper to perform the specified action.

        :param action: A string indicating the action ('open' or 'close').
        """
        if action == 'close':
            # Code to close the gripper
            print("Gripper closed.")
        elif action == 'open':
            # Code to open the gripper
            print("Gripper opened.")
        else:
            print("Invalid gripper action. Use 'open' or 'close'.")

    def sort_objects(self, detected_objects):
        """
        Sort the detected objects based on their properties and control the robotic arm to pick and place them.

        :param detected_objects: A list of detected objects with their bounding box information.
        """
        for obj in detected_objects:
            if obj['type'] == 'color':
                # Define the target position for placing the object based on its color
                target_position = self.get_target_position(obj['color'])
                self.pick_object(obj['bbox'])
                self.place_object(target_position)

    def get_target_position(self, color):
        """
        Define the target position for placing objects based on their color.

        :param color: The color of the object.
        :return: A list containing the target position (x, y, z).
        """
        # Example target positions based on color
        positions = {
            'red': [500, 0, 0],
            'green': [500, 200, 0],
            'blue': [500, 400, 0],
            # Add more colors and their corresponding positions as needed
        }
        return positions.get(color, [500, 0, 0])  # Default position if color not found