tests/test_robot_control.py
import unittest
from robodk import robolink  # Correct import
from src.robot_control import RobotController

class TestRobotController(unittest.TestCase):
    def setUp(self):
        """Set up the RoboDK connection and RobotController instance."""
        self.RDK = robolink.Robolink()  # Use robolink.Robolink()
        self.robot_item = self.RDK.Item('Robot')  # Replace with your robot's name
        self.robot_controller = RobotController(self.robot_item)

    def test_move_to_position(self):
        """Test if the robot can move to a specified position."""
        target_position = [100, 200, 300]
        self.robot_controller.move_to_position(target_position)
        # Add assertions to check if the robot reached the target position

    def test_pick_object(self):
        """Test the pick action of the robotic arm."""
        bbox = (100, 100, 50, 50)
        self.robot_controller.pick_object(bbox)
        # Add assertions to check if the pick action was performed correctly

    def test_place_object(self):
        """Test the place action of the robotic arm."""
        target_position = [400, 400, 0]
        self.robot_controller.place_object(target_position)
        # Add assertions to check if the place action was performed correctly

    def test_sort_objects(self):
        """Test the sorting logic of the robotic arm."""
        detected_objects = [
            {'type': 'color', 'color': 'red', 'bbox': (50, 50, 30, 30)},
            {'type': 'shape', 'shape': 'circle', 'bbox': (150, 150, 40, 40)}
        ]
        sorting_actions = self.robot_controller.sort_objects(detected_objects)
        # Add assertions to check if the sorting actions are correct

if __name__ == '__main__':
    unittest.main()